+++ packages/ui/tests/unit/jest.config.ts
@@ -4,10 +4,10 @@ import { getTransformIgnorePatterns } from '@ttoss/test-utils';
 const config = jestUnitConfig({
   coverageThreshold: {
     global: {
-      statements: 83.5,
-      branches: 79.48,
-      lines: 83.5,
-      functions: 77.36,
+      statements: 84.15,
+      branches: 80,
+      lines: 84.15,
+      functions: 77.4,
     },
   },
   setupFilesAfterEnv: ['./setupTests.tsx'],